+++ internal/api/v1/registration/handler.go
@@ -1,0 +1,19 @@
+package registration
+
+import (
+	"net/http"
+)
+
+type Handler struct {
+	storage storageRepository
+}
+
+func New(storage storageRepository) *Handler {
+	return &Handler{
+		storage: storage,
+	}
+}
+
+func (h *Handler) ServeHTTP(resp http.ResponseWriter, req *http.Request) {
+	defer req.Body.Close()
+}