+++ internal/api/v1/registration/storage.go
@@ -1,0 +1,10 @@
+package registration
+
+import (
+	"context"
+	"github.com/Alexander1000/service-auth/internal/model"
+)
+
+type storageRepository interface {
+	Registration(ctx context.Context, userID int64, pass string, credentials []model.Credential) error
+}